requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
diffusers==0.37.0
et_xmlfile==2.0.0
fastapi==0.128.0
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
importlib_metadata==8.7.1
invisible-watermark==0.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-python==4.13.0.92
openpyxl==3.1.5
openvino==2026.0.0
openvino-telemetry==2025.2.0
openvino-tokenizers==2026.0.0.0
packaging==26.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
protobuf==7.34.0
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydot==3.0.4
Pygments==2.19.2
pyparsing==3.3.2
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
starlette==0.50.0
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
zipp==3.23.0
